Add API to change flow units - not done yet
Added ENsetflowunits. EPANET converts all units to CFS so there is no
need to convert again. The only thing that is not converted is the
Curves data. Currently  the Curve Type is not used by EPANET so it is
difficult to know which type is the curve and how to convert.