Add analytical formula for efficiency calculation
Add number of pumps in the pump group to calculate efficiency of
a pump group. Provide an analytical formula to calculate the
efficiency if information provided as efficiency curve points
describe peak efficiency, peak efficiency flow and cutoff flow
Formula from Ulanicki, Kahler, Coulbeck "Modeling the Efficiency
and Power Characteristics of a Pump Group", Journal of Water
Resources Planning and Management, vol.134, no.1, pp.88-93
Feb. 2008 was used to calculate the parameters of third degree
polyomial

Sorting function sortcurvepoints was added to sort the efficiency
curve points in ascending order with regards to X (flow) in order
to reduce the number of variants to be checked during efficiency
calculation

The efficiency is calculated with function getefficiency if the
number of points provided in the efficiency curve is < 3 or equal
to 3 but with first point = (0,0)